jimt/LABS-472 - busted newly unearthed AssetManager -> AudioRecorder circle
