+++ apps/src/code-studio/components/AssetManager.jsx
@@ -16,29 +16,11 @@ import i18n from '@cdo/locale';
 import {STARTER_ASSET_PREFIX} from '@cdo/apps/assetManagement/assetPrefix';
 import {RecordingFileType} from './recorders';
 
-export const AudioErrorType = {
-  NONE: 'none',
-  INITIALIZE: 'initialize',
-  SAVE: 'save',
-};
-export const ImageMode = {
-  FILE: 'file',
-  ICON: 'icon',
-  URL: 'url',
-  DEFAULT: 'default',
-};
-
-const errorMessages = {
-  403: 'Quota exceeded. Please delete some files and try again.',
-  413: 'The file is too large.',
-  415: 'This type of file is not supported.',
-  500: 'The server responded with an error.',
-  unknown: 'An unknown error occurred.',
-};
-
-const errorUploadDisabled =
-  'This project has been reported for abusive content, ' +
-  'so uploading new assets is disabled.';
+import {
+  AudioErrorType,
+  errorMessages,
+  errorUploadDisabled,
+} from './types/asset-manager';
 
 function getErrorMessage(status) {
   return errorMessages[status] || errorMessages.unknown;
